+++ tests/fast/backends/training_utils/loss/test_loss_snapshot.py
@@ -20,7 +20,7 @@ import torch
 
 from miles.backends.training_utils.loss import compute_advantages_and_returns, loss_function
 from miles.backends.training_utils.loss_hub.corrections import icepop_function, vanilla_tis_function
-from miles.backends.training_utils.loss_hub.logits import get_log_probs_and_entropy, get_values
+from miles.backends.training_utils.loss_hub.logit_processors import get_log_probs_and_entropy, get_values
 from miles.backends.training_utils.loss_hub.losses import policy_loss_function, sft_loss_function, value_loss_function
 
 from .loss_test_utils import (
